使用Int替换所有使用的NSInteger

+++ SwiftVerify/SwiftVerify/Examples/Controller/IM即时通讯/WYTestChatController.swift
@@ -130,7 +130,7 @@ extension WYTestChatController: WYChatViewDelegate {
             replyMessage.lastMessageTimestamp = replyMessage.sharedLastMessageTimestamp(self.chatView.dataSource)
             replyMessage.clientTimestamp = String.wy_sharedDeviceTimestamp()
             replyMessage.sendor = self.sharedUaerInfo(id: "88888", name: "大官人", avatar: "https://img1.baidu.com/it/u=3709586903,1286591012&fm=253&fmt=auto&app=138&f=JPEG")
-            replyMessage.content.text = ["这是自动模拟的回复消息😄😄😄😄😄😄", "回复消息", "这是自动模拟的多行回复消息😄😄😄😄😄😄😄😄😄😄😄😄😄😄😄😄"][NSInteger.wy_random(minimux: 0, maximum: 2)]
+            replyMessage.content.text = ["这是自动模拟的回复消息😄😄😄😄😄😄", "回复消息", "这是自动模拟的多行回复消息😄😄😄😄😄😄😄😄😄😄😄😄😄😄😄😄"][Int.wy_random(minimux: 0, maximum: 2)]
             replyMessage.index = message.index + 1
             replyMessage.sendState = .success
             replyMessage.readers = "1"
@@ -159,7 +159,7 @@ extension WYTestChatController: WYChatViewDelegate {
     }
     
     /// 点击了More控件内某个item
-    func didClickMoreView(_ moreView: WYChatMoreView, _ itemIndex: NSInteger) {
+    func didClickMoreView(_ moreView: WYChatMoreView, _ itemIndex: Int) {
         //wy_print("点击More控件 \(moreView) 内第 \(itemIndex) 个item")
     }
 }
@@ -245,7 +245,7 @@ extension WYTestChatController: WYChatViewEventsHandler {
     }
 
     /// 是否需要内部处理More控件内 cell 的点击事件
-    func canManagerMoreViewClickEvents(_ moreView: WYChatMoreView, _ itemIndex: NSInteger) -> Bool {
+    func canManagerMoreViewClickEvents(_ moreView: WYChatMoreView, _ itemIndex: Int) -> Bool {
         //wy_print("是否需要内部处理More控件内 cell 的点击事件, moreView = \(moreView), itemIndex = \(itemIndex)")
         return true
     }
