+++ SwiftVerify/SwiftVerify/Examples/Controller/瀑布流/WYFlowLayoutAlignmentController.swift
@@ -71,7 +71,7 @@ extension WYFlowLayoutAlignmentController: UICollectionViewDelegate, UICollectio
 
     func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
 //        return 80
-        return NSInteger.wy_random(minimux: 1, maximum: 129)
+        return Int.wy_random(minimux: 1, maximum: 129)
 //        return 9
 //        return 169
     }
@@ -87,7 +87,7 @@ extension WYFlowLayoutAlignmentController: UICollectionViewDelegate, UICollectio
                     if wy_collectionView(collectionView, layout: collectionViewLayout, hoverForHeaderForSectionAt: indexPath.section) == true {
                         return CGSize(width: 35, height: indexPath.row + 10)
                     }else {
-                        return CGSize(width: 35, height: NSInteger.wy_random(minimux: 35, maximum: 135))
+                        return CGSize(width: 35, height: Int.wy_random(minimux: 35, maximum: 135))
                     }
                 }
             }else {
@@ -133,7 +133,7 @@ extension WYFlowLayoutAlignmentController: UICollectionViewDelegate, UICollectio
             if collectionView.isPagingEnabled == true {
                 return CGSize(width: 35, height: 35)
             }
-            return CGSize(width: NSInteger.wy_random(minimux: 35, maximum: 135), height: 35)
+            return CGSize(width: Int.wy_random(minimux: 35, maximum: 135), height: 35)
         }
     }
     
@@ -153,7 +153,7 @@ extension WYFlowLayoutAlignmentController: UICollectionViewDelegate, UICollectio
         return 10
     }
     
-    func wy_collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, itemNumberOfLinesForSectionAt section: NSInteger) -> NSInteger {
+    func wy_collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, itemNumberOfLinesForSectionAt section: Int) -> Int {
         return 3
     }
     
@@ -172,25 +172,25 @@ extension WYFlowLayoutAlignmentController: UICollectionViewDelegate, UICollectio
         return true
     }
     
-    func wy_collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, hoverForHeaderForSectionAt section: NSInteger) -> Bool {
+    func wy_collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, hoverForHeaderForSectionAt section: Int) -> Bool {
         if wy_collectionView(collectionView, layout: collectionViewLayout, flowLayoutAlignmentForSectionAt: section) != .default {
             return false
         }
         return true
     }
     
-    func wy_collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, spacingBetweenHeaderAndLastPartitionFooter section: NSInteger) -> CGFloat {
+    func wy_collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, spacingBetweenHeaderAndLastPartitionFooter section: Int) -> CGFloat {
         return CGFloat(section) * 10
     }
     
-    func wy_collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, referenceSizeForHeaderInSection section: NSInteger) -> CGSize {
+    func wy_collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, referenceSizeForHeaderInSection section: Int) -> CGSize {
         if collectionView.isPagingEnabled || collectionView == horizontal {
             return .zero
         }
         return CGSize(width: UIDevice.wy_screenWidth, height: 50)
     }
     
-    func wy_collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, referenceSizeForFooterInSection section: NSInteger) -> CGSize {
+    func wy_collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, referenceSizeForFooterInSection section: Int) -> CGSize {
         if collectionView.isPagingEnabled || collectionView == horizontal {
             return .zero
         }
