+++ WYBasisKit/WYBasisKit/WYBasisKit/Layout/ChatView/Views/Handler/WYChatViewDelegateHandler.swift
@@ -51,7 +51,7 @@ import UIKit
     @objc optional func canManagerEmojiSendViewClickEvents(_ sendView: UIButton) -> Bool
     
     /// 是否需要内部处理More控件内 cell 的点击事件
-    @objc optional func canManagerMoreViewClickEvents(_ moreView: WYChatMoreView, _ itemIndex: NSInteger) -> Bool
+    @objc optional func canManagerMoreViewClickEvents(_ moreView: WYChatMoreView, _ itemIndex: Int) -> Bool
     
     /// 是否需要内部处理tableView代理 cellForRowAt 方法
     @objc optional func canManagerCellForRowEvents(_ chatView: WYChatView, _ tableView: UITableView, _ indexPath: IndexPath) -> UITableViewCell?
@@ -99,5 +99,5 @@ import UIKit
     @objc optional func emojiItemLongPress(_ gestureRecognizer: UILongPressGestureRecognizer, emoji: String, imageView: UIImageView)
     
     /// 点击了More控件内某个item
-    @objc optional func didClickMoreView(_ moreView: WYChatMoreView, _ itemIndex: NSInteger)
+    @objc optional func didClickMoreView(_ moreView: WYChatMoreView, _ itemIndex: Int)
 }
